commit 25/02/2023

+++ application/models/Servicemodel.php
@@ -15,17 +15,18 @@ class Servicemodel extends CI_Model {
     var $table_item = 'items';
     var $table_payment = 'payment_methods';
     var $table_claim = 'claim_paids';
-    var $column_order = array(null, 'transaction_date', 'customer_id', 'code', 'total', 'cash', 'changes', 'method_id', 'is_cash', 'status', 'type', 'note');
-    var $column_search = array('s.code', 'c.name');
+    var $table_stock = 'stocks';
+    
+    var $column_order = array(null, 'transaction_date', 'is_customer', 'customer_id', 'customer_name', 'code', 'total', 'cash', 'changes', 'method_id', 'is_cash', 'status', 'type', 'note');
+    var $column_search = array('s.code', 's.customer_name');
     var $order = array('s.id' => 'desc');
 
     private function _getDatatablesQuery()
     {
-        $this->db->select('*, s.id as s_id, c.name as c_name');
+        $this->db->select('*');
         $this->db->where('s.deleted_at', null);
         $this->db->where('s.type', 'service');
         $this->db->from('sales as s');
-        $this->db->join('customers as c', 'c.id = s.customer_id');
  
         $i = 0;
      
@@ -89,7 +90,24 @@ class Servicemodel extends CI_Model {
     {
         $this->db->order_by('name', 'asc');
         $this->db->where('deleted_at', null);
+        $this->db->where('stock >', 0);
         $query = $this->db->get($this->table_item);
+        if ($query->num_rows() > 0) {
+            return $query->result_array();
+        }
+        else {
+            return array();
+        }
+    }
+
+    public function searchItems($keyword, $offset, $limit, $uid)
+    {
+        $this->db->order_by('name', 'asc');
+        $this->db->where('deleted_at', null);
+        // $this->db->where('stock >', 0);
+        $this->db->where_not_in('id', $uid);
+        $this->db->like('name', $keyword);
+        $query = $this->db->get($this->table_item, $limit, $offset);
         if ($query->num_rows() > 0) {
             return $query->result_array();
         }
@@ -113,6 +131,7 @@ class Servicemodel extends CI_Model {
 
     public function getDataDetailByIdTrans($id)
     {
+        $this->db->select('sale_details.id, sale_details.sale_id, sale_details.item_id, sale_details.qty, sale_details.price, i.name, i.stock');
         $this->db->where('sale_id', $id);
         $this->db->join('items as i', 'i.id = sale_details.item_id');
         $query = $this->db->get($this->table_detail);
@@ -128,13 +147,17 @@ class Servicemodel extends CI_Model {
     {
         $this->db->trans_begin();
         $dataInsert = array('transaction_date' => $data['transaction_date'],
+                        'is_customer'  => $data['is_customer'],
                         'customer_id'  => $data['customer_id'],
+                        'customer_name'  => $data['customer_name'],
                         'code'  => '-',
                         'total'  => $data['total'],
                         'cash'  => $data['cash'],
                         'changes' => $data['changes'],
                         'method_id' => $data['method_id'],
+                        'type_service' => $data['type_service'],
                         'is_cash'  => $data['is_cash'],
+                        'payment_at' => ($data['is_cash'] == 0) ? null : date('Y-m-d H:i:s'),
                         'status' => 2,
                         'type' => 'service',
                         'note' => $data['note'],
@@ -144,27 +167,63 @@ class Servicemodel extends CI_Model {
         $this->db->insert($this->table,$dataInsert);
         $idx = $this->db->insert_id();
 
-        foreach ($data['item_ids'] as $key => $value) {
-            $dataInsertDetail = array('sale_id' => $idx,
+        if (!empty($data['item_ids'])) {
+            foreach ($data['item_ids'] as $key => $value) {
+                $dataInsertDetail = array('sale_id' => $idx,
                                     'item_id' => $value,
-                                    'qty' => $data['qty'][$key]
+                                    'qty' => $data['qty'][$key],
+                                    'price' => $data['price'][$key]
                                     );
-            $this->db->insert($this->table_detail, $dataInsertDetail);
-            $dataInsertMutation = array('transaction_id' => $idx,
-                                    'item_id' => $value,
-                                    'amount' => '-'.$data['qty'][$key],
-                                    'type' => 'service',
-                                    'created_at' => date('Y-m-d H:i:s')
-                                    );
-            $this->db->insert('stock_mutations', $dataInsertMutation);
+                $this->db->insert($this->table_detail, $dataInsertDetail);
 
-            $res = $this->decreaseStock($value, $data['qty'][$key]);
+                $dataInsertMutation = array('transaction_id' => $idx,
+                                        'item_id' => $value,
+                                        'amount' => '-'.$data['qty'][$key],
+                                        'type' => 'sale',
+                                        'created_at' => date('Y-m-d H:i:s')
+                                        );
+                $this->db->insert('stock_mutations', $dataInsertMutation);
 
-            if ($res == 0) {
-                $this->db->trans_rollback();
-                return array('msg' => 'fail');
+                $this->db->where('deleted_at', null);
+                $this->db->where('item_id', $value);
+                $this->db->where('sale_id IS NULL', null, false);
+                $row = $this->db->get($this->table_stock)->num_rows();
+                if ($row < $data['qty'][$key]) {
+                    $this->db->trans_rollback();
+                    return array('msg' => 'fail');
+                }
+                else {
+                    $limit = $data['qty'][$key];
+                    $query = "UPDATE stocks SET sale_id='$idx'
+                                WHERE id IN (
+                                    SELECT id FROM (
+                                        SELECT id FROM stocks
+                                        WHERE sale_id IS NULL
+                                        AND
+                                        item_id = '$value'
+                                        AND
+                                        deleted_at IS null
+                                        ORDER BY id ASC
+                                        LIMIT 0, $limit
+                                    ) tmp
+                                )";
+                    $this->db->query($query);
+                    if ($this->db->affected_rows() == 0) {
+                        $this->db->trans_rollback();
+                        return array('msg' => 'fail');
+                    }
+                }
+
+                $res = $this->decreaseStock($value, $data['qty'][$key]);
+
+                if ($res == 0) {
+                    $this->db->trans_rollback();
+                    return array('msg' => 'fail');
+            }
             }
         }
+
+
         $this->db->trans_commit();
         return array('msg' => 'success', 'trans_id' => $idx);
     }
@@ -189,12 +248,14 @@ class Servicemodel extends CI_Model {
     {
         $this->db->trans_begin();
         $dataUpdate = array('transaction_date' => $data['transaction_date'],
+                        'is_customer'  => $data['is_customer'],
                         'customer_id'  => $data['customer_id'],
-                        'code'  => '-',
+                        'customer_name'  => $data['customer_name'],
                         'total'  => $data['total'],
                         'cash'  => $data['cash'],
                         'changes' => $data['changes'],
                         'method_id' => $data['method_id'],
+                        'type_service' => $data['type_service'],
                         'is_cash'  => $data['is_cash'],
                         'status' => 2,
                         'type' => 'service',
@@ -205,32 +266,69 @@ class Servicemodel extends CI_Model {
         $this->db->where('id', $id);
         $this->db->update($this->table, $dataUpdate);
 
-        $this->db->where('sale_id', $id);
-        $this->db->delete($this->table_detail);
+        if (!empty($data['item_ids'])) {
+            $this->db->where('sale_id', $id);
+            $this->db->update($this->table_stock, array('sale_id' => null));
 
-        $this->db->where('transaction_id', $id);
-        $this->db->where('type', 'service');
-        $this->db->delete('stock_mutations');
+            $this->db->where('sale_id', $id);
+            $this->db->delete($this->table_detail);
 
-        foreach ($data['item_ids'] as $key => $value) {
-            $dataInsertDetail = array('sale_id' => $id,
+            $this->db->where('transaction_id', $id);
+            $this->db->where('type', 'service');
+            $this->db->delete('stock_mutations');
+
+            foreach ($data['item_ids'] as $key => $value) {
+                $dataInsertDetail = array('sale_id' => $id,
                                     'item_id' => $value,
-                                    'qty' => $data['qty'][$key]
+                                    'qty' => $data['qty'][$key],
+                                    'price' => $data['price'][$key]
                                     );
-            $this->db->insert($this->table_detail, $dataInsertDetail);
-            $dataInsertMutation = array('transaction_id' => $id,
-                                    'item_id' => $value,
-                                    'amount' => '-'.$data['qty'][$key],
-                                    'type' => 'service',
-                                    'created_at' => date('Y-m-d H:i:s')
-                                    );
-            $this->db->insert('stock_mutations', $dataInsertMutation);
+                $this->db->insert($this->table_detail, $dataInsertDetail);
 
-            $res = $this->decreaseStock($value, $data['qty'][$key]);
+                $dataInsertMutation = array('transaction_id' => $id,
+                                        'item_id' => $value,
+                                        'amount' => '-'.$data['qty'][$key],
+                                        'type' => 'sale',
+                                        'created_at' => date('Y-m-d H:i:s')
+                                        );
+                $this->db->insert('stock_mutations', $dataInsertMutation);
 
-            if ($res == 0) {
-                $this->db->trans_rollback();
-                return array('msg' => 'fail');
+                $this->db->where('deleted_at', null);
+                $this->db->where('item_id', $value);
+                $this->db->where('sale_id IS NULL', null, false);
+                $row = $this->db->get($this->table_stock)->num_rows();
+                if ($row < $data['qty'][$key]) {
+                    $this->db->trans_rollback();
+                    return array('msg' => 'fail');
+                }
+                else {
+                    $limit = $data['qty'][$key];
+                    $query = "UPDATE stocks SET sale_id='$id'
+                                WHERE id IN (
+                                    SELECT id FROM (
+                                        SELECT id FROM stocks
+                                        WHERE sale_id IS NULL
+                                        AND
+                                        item_id = '$value'
+                                        AND
+                                        deleted_at IS null
+                                        ORDER BY id ASC
+                                        LIMIT 0, $limit
+                                    ) tmp
+                                )";
+                    $this->db->query($query);
+                    if ($this->db->affected_rows() == 0) {
+                        $this->db->trans_rollback();
+                        return array('msg' => 'fail');
+                    }
+                }
+
+                $res = $this->decreaseStock($value, $data['qty'][$key]);
+
+                if ($res == 0) {
+                    $this->db->trans_rollback();
+                    return array('msg' => 'fail');
+                }
             }
         }
         $this->db->trans_commit();
@@ -249,6 +347,9 @@ class Servicemodel extends CI_Model {
         $data = array('deleted_at' => date('Y-m-d H:i:s'));
         $this->db->where('id',$id);
         $this->db->update($this->table,$data);
+
+        $this->db->where('sale_id', $id);
+        $this->db->update($this->table_stock, array('sale_id' => null));
 
         $this->db->where('sale_id', $id);
         $this->db->delete($this->table_detail);
